requirements.txt
flask
ultralytics
torch
torchvision
pillow
opencv-python-headless
gunicorn
gdown
